DependencyInjection: ObservationLocationRepository
