+++ Mage/Repository/RepositoryManager.swift
@@ -15,5 +15,4 @@ class RepositoryManager: ObservableObject {
     var observationIconRepository: ObservationIconRepository?
     var observationsTileRepository: ObservationsTileRepository?
     var observationsMapFeatureRepository: ObservationsMapFeatureRepository?
-    var observationLocationRepository: ObservationLocationRepository?
 }
